Protect vs resource leakage
Previously, if a semaphore was given a block and then went stale, but
eventually completed it's task, it would add a duplicate of its' token
to the available pool. Additionally, any codebase that only used the
block structure would not self-heal from this state.

Non-block usage would self-heal thanks to the unlock method.